dbcrapi/dbcr_project/requirements.txt
boto3==1.12.41
botocore==1.15.41
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.38.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.9.2
docutils==0.15.2
gunicorn==20.0.4
idna==2.8
jmespath==0.9.5
oauthlib==3.1.0
Pillow==7.1.1
psycopg2==2.8.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
